==> gradient_networks/terrestrial/04_gradient_diff_from_random.Rmd <==
---
title: "04_gradient_diff_from_random"
output: html_document
date: "2023-08-22"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
fung_data <- readRDS("~/cmaiki_lts/kaciekaj/waimea/intermediates/gradient/terrestrial/six_empos_fung_grad_data_ordered.rds")
bact_data <- readRDS("~/cmaiki_lts/kaciekaj/waimea/intermediates/gradient/terrestrial/six_empos_bact_grad_data_ordered.rds")


remove_dummy_row <- function(df) {
  new_df <- df[row.names(df)!="dummy",]
  return(new_df)
}


fung_data <- lapply(fung_data, remove_dummy_row)
bact_data <- lapply(bact_data, remove_dummy_row)

# import igraph objects
fung_igraphs <- readRDS("../../intermediates/gradient/terrestrial/fung_gradient_igraphs.rds")
bact_igraphs <- readRDS("../../intermediates/gradient/terrestrial/bact_gradient_igraphs.rds")
cross_igraphs <- readRDS("../../intermediates/gradient/terrestrial/cross_gradient_igraphs.rds")

# metadata
fung_meta_terr <- readRDS("../../intermediates/gradient/terrestrial/fung_terrestrial_metadata.rds")
bact_meta_terr <- readRDS("../../intermediates/gradient/terrestrial/bact_terrestrial_metadata.rds")

plots <- unique(fung_meta_terr$site_name)
```

```{r}
fung_power_law <- lapply(fung_igraphs, function(x) power.law.fit(degree(x))$KS.p)
fung_power_law > 0.05

bact_power_law <- lapply(bact_igraphs, function(x) power.law.fit(degree(x))$KS.p)
bact_power_law > 0.05

cross_power_law <- lapply(cross_igraphs, function(x) power.law.fit(degree(x))$KS.p)
cross_power_law > 0.05


tab <- data.frame("Fungi" = unlist(fung_power_law),
                  "Bacteria" = unlist(bact_power_law),
                  "Cross-domain" = unlist(cross_power_law))
rownames(tab) <- plots

tab <- tab[c("Beach", "Estuary", "Entrance", "Confluence", "Waterfall", "DrumRoad", "Ridge"),]

write.csv(tab, "../../outputs/gradient_scalefree_table.csv")


# small world
library(qgraph)

fung_sw = lapply(fung_igraphs, smallworldIndex)
bact_sw <- lapply(bact_igraphs, smallworldIndex)
cross_sw <- lapply(cross_igraphs, smallworldIndex)

sw_vals <- data.frame(Fungi = sapply(fung_sw, function(x) x$index),
                      Bacteria = sapply(bact_sw, function(x) x$index),
                      Cross = sapply(cross_sw, function(x) x$index))

rownames(sw_vals) <- plots

sw_vals <- sw_vals[c("Beach", "Estuary", "Entrance", "Confluence", "Waterfall", "DrumRoad", "Ridge"),]

write.csv(sw_vals, "../../outputs/gradient_smallworld_table.csv")

```

## Abundance vs degree

```{r}
# regressions

### functions
# add reads column to degree df
add_reads <- function (deg_df, otu_table) {
  deg_df$reads <- rowSums(otu_table)
  return(deg_df)
}

#define function to extract overall p-value of model
overall_p <- function(my_model) {
    f <- summary(my_model)$fstatistic
    p <- pf(f[1],f[2],f[3],lower.tail=F)
    attributes(p) <- NULL
    return(p)
}


### fungi
fung_deg <- lapply(fung_igraphs, function(x) data.frame(degree = degree(x)))

all(rownames(fung_deg[[1]]) == rownames(fung_data[[1]]))

fung_dfs <- mapply(add_reads,
               fung_deg,
               fung_data,
               SIMPLIFY = FALSE)

fung_lms <- lapply(fung_dfs, function(x) lm(log(reads)~sqrt(degree+1), data = x))

fung_r2 <- sapply(fung_lms, function(x) round(summary(x)$r.squared, 4))
fung_p <- sapply(fung_lms, function(x) round(overall_p(x), 4))

fung_reg_res <- data.frame(type = rep("Fungi", 7),
                           plot = plots,
                           r2 = fung_r2,
                           p = fung_p)
fung_reg_res$plot <- factor(fung_reg_res$plot, levels=c("Beach", "Estuary", "Entrance", "Confluence", "Waterfall", "DrumRoad", "Ridge"))
fung_reg_res <- fung_reg_res[order(fung_reg_res$plot),]


### bacteria
bact_deg <- lapply(bact_igraphs, function(x) data.frame(degree = degree(x)))

all(rownames(bact_deg[[1]]) == rownames(bact_data[[1]]))

#test_glm <- glm(log(reads)~sqrt(degree), family = "quasipoisson", data = bact_dfs[[1]])

# resid(bact_lms[[1]]) #List of residuals
# plot(density(resid(bact_lms[[1]]))) #A density plot
# qqnorm(resid(m1)) # A quantile normal plot - good for checking normality
# qqline(resid(m1))


bact_dfs <- mapply(add_reads,
               bact_deg,
               bact_data,
               SIMPLIFY = FALSE)

bact_lms <- lapply(bact_dfs, function(x) lm(log(reads)~sqrt(degree), data = x))

bact_r2 <- sapply(bact_lms, function(x) round(summary(x)$r.squared, 4))
bact_p <- sapply(bact_lms, function(x) round(overall_p(x), 4))

bact_reg_res <- data.frame(type = rep("Bacteria", 7),
                           plot = plots,
                           r2 = bact_r2,
                           p = bact_p)
bact_reg_res$plot <- factor(bact_reg_res$plot, levels=c("Beach", "Estuary", "Entrance", "Confluence", "Waterfall", "DrumRoad", "Ridge"))
bact_reg_res <- bact_reg_res[order(bact_reg_res$plot),]


### cross
cross_deg <- lapply(cross_igraphs, function(x) data.frame(degree = degree(x)))

all(names(fung_data[[1]])==names(bact_data[[1]]))

cross_data <- mapply(rbind, fung_data, bact_data, SIMPLIFY = FALSE)

all(rownames(cross_deg[[1]]) == rownames(cross_data[[1]]))

cross_dfs <- mapply(add_reads,
               cross_deg,
               cross_data,
               SIMPLIFY = FALSE)

cross_lms <- lapply(cross_dfs, function(x) lm(log(reads)~sqrt(degree), data = x))

cross_r2 <- sapply(cross_lms, function(x) round(summary(x)$r.squared, 4))
cross_p <- sapply(cross_lms, function(x) round(overall_p(x), 4))

cross_reg_res <- data.frame(type = rep("Cross", 7),
                           plot = plots,
                           r2 = cross_r2,
                           p = cross_p)

cross_reg_res$plot <- factor(cross_reg_res$plot, levels=c("Beach", "Estuary", "Entrance", "Confluence", "Waterfall", "DrumRoad", "Ridge"))
cross_reg_res <- cross_reg_res[order(cross_reg_res$plot),]


all_data <- do.call("rbind", list(fung_reg_res, bact_reg_res, cross_reg_res))
names(all_data) <- c("Network Type", "Plot", "R2", "P")

write.csv(all_data, "../../outputs/gradient_abun_vs_degree_regression_values.csv")
```
